fix string to uint conversion, simplify expressions

+++ 03/main.cpp
@@ -8,12 +8,12 @@
 
 int main()
 {
-  int n = 0;
+  double n = 0;
 
   Matrix m1(2, 3);
-  for (int row = 0; row < 2; ++row)
+  for (double row = 0; row < 2; ++row)
   {
-    for (int col = 0; col < 3; ++col)
+    for (double col = 0; col < 3; ++col)
     {
       m1[row][col] = ++n;
     }
